+++ src/GlobalStyle.js
@@ -1,0 +1,14 @@
+import { createGlobalStyle } from "styled-components";
+
+export const GlobalStyles = createGlobalStyle`
+    *{
+        margin: 0;
+        padding: 0;
+    }
+
+    body{
+        font-family: 'Roboto', sans-serif;
+    }
+
+
+`
