+++ src/components/Pokecard/Pokecard.jsx
@@ -1,0 +1,61 @@
+import { useState } from "react";
+import { useEffect } from "react";
+import React from "react";
+import axios from "axios";
+import pokebolaTransp from "../../assets/pokebolaTransp.png"
+// import pokebola from "../../assets/pokebola.png"
+import {
+  CardContainer,
+  CardFather,
+  Imagem,
+  CaptureButton, 
+  Stats
+} from "./PokeStyles";
+
+const PokeCard = (props) => {
+  const [pokemon, setPokemon] = useState([]);
+
+  useEffect(() => {
+
+    pegaPokemon(props.pokemon);
+
+  }, [props.pokemon]);
+
+  const pegaPokemon = () => {
+    axios
+      .get(`https://pokeapi.co/api/v2/pokemon/${props.nome}`)
+      .then((response) => {
+        console.log(pokemon)
+        setPokemon(response.data);
+        console.log(response.data)
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  };
+
+  return (
+    <CardContainer>
+      <CardFather>
+        <h3>{pokemon.name && <>{pokemon.name.toUpperCase()}</>}</h3>
+        <Stats>
+          <p>{pokemon.types && <>{pokemon.types[0].type.name}</>}</p>
+          <p>{pokemon.types && (pokemon.types.length > 1 ? <>|{pokemon.types[1].type.name}</> : "")}</p>
+          {/* <img src={pokebola} height={25} onClick={console.log('clicou')}/> */}
+          {/* <button img={}>Clique</button> */}
+        </Stats>
+        {/* <img src={pokebolaTransp} height={25}/> */}
+          <Imagem>        
+          {pokemon.sprites && (
+            <img src={pokemon.sprites.other.home.front_default} alt={pokemon.name} />
+          )}
+        </Imagem>
+        <CaptureButton>
+        <img src={pokebolaTransp} />
+        </CaptureButton>
+      </CardFather>
+    </CardContainer>
+  );
+};
+
+export default PokeCard;
