+++ src/global/GlobalContext.js
@@ -1,0 +1,3 @@
+import React from "react";
+
+export const GlobalContext = React.createContext();
