+++ src/global/GlobalState.js
@@ -1,0 +1,6 @@
+// import React, { useState, useEffect } from "react";
+// import { BASE_URL } from "../Constants/BASE_URL";
+// import { GlobalContext } from "./GlobalContext";
+// import axios from "axios";
+
+
