--- src/index.css
@@ -1,9 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Encode+Sans+SC:wght@300&display=swap');
-
-*{
-  padding: 0;
-  margin: 0;
-
-
-}
-
